Work around missing host repositories
When the host profile references an overlay which is not used
within the target, emerge fails to locate its profile correctly.
We avoid that by allowing the variable EBOOTSTRAP_PROFILE_REPOS
to be set in the configuration file. Any repos referenced there
will be appended to PORTAGE_REPOSITORIES provided to emerge in
ebootstrap-emerge, if they don't already exist in the target
configuration.

This is a workaround solution and might cause unwanted side
effects with dependency resolution.