Remove the merge action
Merge is used by ebuild when installing a package into the live filesystem.
We never want to do this with an eroot, so it should no be default action.

This means, however, that there is no way to get the config action to be
run after an install is complete. This is just a consequence of the
mismatch between the ebuild phases and ebootstrap phases (and a good
example of why the ebuild backend is probably a bad idea).