Ensure the host profile repo is available when running emerge with --config-root
All of the repos used by the host's profile must be defined in the target
portage config, otherwise emerge commplains about a missing profile.

This change enumerates through the host profile parents and adds any repos
which are used by the host but not the target to the PORTAGE_REPOSITORIES
config used when running emerge for the target ebootstrap root path.