Updates prepare to mount paths based on the current target portage config
Uses the config extracted from EROOT/etc/portage, rather than E_REPOS,
E_PKGDIR and E_DISTDIR. This makes the routine more flexible, because it is
not so dependent on the eroot configuration directly.

Also required that ebootstrap-configure-portage is called during the prepare
processing, so that the config is available.