Reimplements the ebootstrap-configure-make-conf function
This fixes the idempotency issue with the generation of make.conf by
parsing the contents of E_MAKE_CONF from the config file and updating
the applicable variables in make.conf, rather than simply appending to
the end of the file.

It also lines up the default make.conf content with what is in the
current stage3 created by catalyst.