Fixes configure error when /etc/portage does not exist
When installing with nostage3, the /etc/portage directory was not
created initially if there a no specific repos defined in the eroot
config. It then failed creating the profile link.

This changes reorders the portage configuration process so that
make.conf is created first. That guarantees /etc/portage will be
created before attempting to set the profile.