Exit early when baselayout install fails
Currently we only warn about error codes when emerge fails. That is
because some errors are non-fatal, and we have no way of knowing which
it is.

Example is the gtk+ binary packages fail a post-install setup, but the
install can continue successfuly regardless. (It might be an idea to
re-do the failed step inside a chrooted emerge, but we don't have
sufficient feedback to handle that currently).