Adds a warning when the default profile does not match the stage3
E_PROFILE does not need to be set; in that case the system default
profile will be used. However, typically it is desirable to set a
non-default profile.

If the stage3 profile is non-default, then the E_PROFILE should always
be set inside the eroot config. Otherwise the binpkgs cache will be
built with a different profile than what the eroot is configured with,
and that's likely to cause errors for nostage3 installs.

This change adds a warning when installing from a stage3 if the
E_PROFILE is undefined and the stage uses a non-default profile.