Fixes installing packages which create users and groups
When installing into an EROOT, users and groups created by packages
are added to the host system, not the EROOT. This then causes problems
when booting into the ebootsrapped system, because the users do not
exist there.

This change searches for packages in the eroot which inherited the
user eclass and re-installs them inside a chrooted emerge command.